azure_search_demo/Backend/requirements.txt
azure-functions
azure-search-documents
requests
